requirements.txt
google-cloud-firestore>=2.1.3
google-auth>=1.33.0
